Make feeEstimateToV0_7 private
