+++ rpc/estimate_fee_test.go
@@ -164,38 +164,3 @@ func assertEqualCairo1Class(t *testing.T, cairo1Class *core.Cairo1Class, class *
 		assert.Equal(t, cairo1Class.EntryPoints.External[idx].Selector, class.EntryPoints.External[idx].Selector)
 	}
 }
-
-func TestFeeEstimateToV0_7(t *testing.T) {
-	t.Run("empty", func(t *testing.T) {
-		assert.Equal(t, rpc.FeeEstimateV0_7{}, rpc.FeeEstimateToV0_7(rpc.FeeEstimate{}))
-	})
-
-	t.Run("full", func(t *testing.T) {
-		gasConsumed := new(felt.Felt).SetUint64(1)
-		gasPrice := new(felt.Felt).SetUint64(2)
-		dataGasConsumed := new(felt.Felt).SetUint64(3)
-		dataGasPrice := new(felt.Felt).SetUint64(4)
-		overallFee := new(felt.Felt).SetUint64(5)
-		unit := rpc.WEI
-		assert.Equal(
-			t,
-			rpc.FeeEstimateV0_7{
-				GasConsumed:     gasConsumed,
-				GasPrice:        gasPrice,
-				DataGasConsumed: dataGasConsumed,
-				DataGasPrice:    dataGasPrice,
-				OverallFee:      overallFee,
-				Unit:            &unit,
-			},
-			rpc.FeeEstimateToV0_7(rpc.FeeEstimate{
-				L1GasConsumed:     gasConsumed,
-				L1GasPrice:        gasPrice,
-				L2GasConsumed:     new(felt.Felt).SetUint64(6),
-				L2GasPrice:        new(felt.Felt).SetUint64(7),
-				L1DataGasConsumed: dataGasConsumed,
-				L1DataGasPrice:    dataGasPrice,
-				OverallFee:        overallFee,
-				Unit:              &unit,
-			}))
-	})
-}
